pyproject.toml
[project]
name = "table-to-text"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.3.21",
    "langchain-community>=0.3.20",
    "langchain-ollama>=0.3.0",
    "loguru>=0.7.3",
    "pandas>=2.2.3",
    "pyyaml>=6.0.2",
]
